requirements.txt
google-generativeai
pillow
opencv-python
numpy
requests
pillow
pydantic
python-dotenv
pandas
openpyxl